requirements.txt
pytelegrambotapi
google-generativeai